requirements-mac.txt
pyobjc==9.0.1
pyotp==2.8.0
pyperclip==1.8.2